Slimmed down IfcStore: factored out significant routines to extension methods on IModel so we can use elsewhere
