+++ Xbim.Ifc/IfcStore.cs
@@ -1,24 +1,20 @@
 ﻿using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
-using System.IO.Compression;
 using System.Linq;
 using System.Text.RegularExpressions;
-using System.Xml;
 using Xbim.Common;
 using Xbim.Common.Exceptions;
 using Xbim.Common.Federation;
 using Xbim.Common.Geometry;
 using Xbim.Common.Metadata;
 using Xbim.Common.Step21;
-using Xbim.Common.XbimExtensions;
+using Xbim.Ifc.Extensions;
 using Xbim.Ifc4.Interfaces;
 using Xbim.Ifc4.MeasureResource;
 using Xbim.IO;
 using Xbim.IO.Step21;
-using Xbim.IO.Xml;
 
 namespace Xbim.Ifc
 {
@@ -34,19 +30,11 @@ namespace Xbim.Ifc
     /// </remarks>
     public class IfcStore : IModel, IDisposable, IFederatedModel, IEquatable<IModel>
     {
-        private IModel _model;
-        private XbimSchemaVersion _schema;
         private const string RefDocument = "XbimReferencedModel";
         public event NewEntityHandler EntityNew;
         public event ModifiedEntityHandler EntityModified;
         public event DeletedEntityHandler EntityDeleted;
 
-        public IInverseCache InverseCache
-        {
-            get { return _model.InverseCache; }
-        }
-
-        public object Tag { get; set; }
 
         private bool _disposed;
 
@@ -88,6 +76,13 @@ namespace Xbim.Ifc
             AssignModel(model, editorDetails, ifcVersion);
         }
 
+        public IModel Model
+        {
+            get;
+            private set;
+        }
+        public XbimEditorCredentials EditorDetails { get; private set; }
+
         /// <summary>
         /// Provides access to model persistance capabilities
         /// </summary>
@@ -105,13 +100,11 @@ namespace Xbim.Ifc
 
         private void AssignModel(IModel model, XbimEditorCredentials editorDetails, XbimSchemaVersion schema)
         {
-            _model = model;
-            _model.EntityNew += _model_EntityNew;
-            _model.EntityDeleted += _model_EntityDeleted;
-            _model.EntityModified += _model_EntityModified;
+            Model = model;
+            Model.EntityNew += Model_EntityNew;
+            Model.EntityDeleted += Model_EntityDeleted;
+            Model.EntityModified += Model_EntityModified;
             FileName = null;
-            _schema = schema;
-
             SetupEditing(editorDetails);
 
             LoadReferenceModels();
@@ -141,27 +134,32 @@ namespace Xbim.Ifc
                 EditorDetails = editorDetails;
             }
 
-            _model.EntityNew += IfcRootInit;
-            _model.EntityModified += IfcRootModified;
+            Model.EntityNew += IfcRootInit;
+            Model.EntityModified += IfcRootModified;
         }
 
 
-        private void _model_EntityDeleted(IPersistEntity entity)
+        /// <summary>
+        /// Creates a Database store at the specified location
+        /// </summary>
+        /// <param name="filePath"></param>
+        /// <param name="editorDetails"></param>
+        /// <param name="ifcVersion"></param>
+        /// <returns></returns>
+        public static IfcStore Create(string filePath, XbimEditorCredentials editorDetails, XbimSchemaVersion ifcVersion)
         {
-            if (EntityDeleted != null) EntityDeleted.Invoke(entity);
+            return new IfcStore(filePath, ifcVersion, editorDetails);
         }
 
-        private void _model_EntityNew(IPersistEntity entity)
+        public static IfcStore Create(XbimEditorCredentials editorDetails, XbimSchemaVersion ifcVersion, XbimStoreType storageType)
         {
-            if (EntityNew != null) EntityNew.Invoke(entity);
+            return new IfcStore(storageType, ifcVersion, editorDetails);
         }
 
-        private void _model_EntityModified(IPersistEntity entity, int property)
+        public static IfcStore Create(XbimSchemaVersion ifcVersion, XbimStoreType storageType)
         {
-            if (EntityModified != null) EntityModified.Invoke(entity, property);
+            return new IfcStore(storageType, ifcVersion, null);
         }
-
-       
         /// <summary>
         /// You can use this function to open IFC model from a <see cref="Stream"/>. 
         /// You need to know file type (IFC, IFCZIP, IFCXML) and schema type (IFC2x3 or IFC4) to be able to use this function.
@@ -180,7 +178,8 @@ namespace Xbim.Ifc
         /// <param name="progDelegate">Progress reporting delegate</param>
         /// <param name="codePageOverride">
         /// A CodePage that will be used to read implicitly encoded one-byte-char strings. If -1 is specified the default ISO8859-1
-        /// encoding will be used accoring to the Ifc specification. </param>/// <returns></returns>
+        /// encoding will be used accoring to the Ifc specification. </param>
+        /// <returns></returns>
         public static IfcStore Open(Stream stream, StorageType dataType, XbimSchemaVersion schema, XbimModelType modelType, XbimEditorCredentials editorDetails = null, 
             XbimDBAccess accessMode = XbimDBAccess.Read, ReportProgressDelegate progDelegate = null, int codePageOverride = -1)
         {
@@ -230,145 +229,119 @@ namespace Xbim.Ifc
 
         }
 
+        #region IModel
+
+        public object Tag { get; set; }
+        public ILogger Logger { get => Model.Logger; set => Model.Logger = value; }
+        public IInverseCache InverseCache
+        {
+            get { return Model.InverseCache; }
+        }
+
         public int UserDefinedId
         {
-            get { return _model.UserDefinedId; }
-            set { _model.UserDefinedId = value; }
+            get { return Model.UserDefinedId; }
+            set { Model.UserDefinedId = value; }
         }
 
         public IGeometryStore GeometryStore
         {
-            get { return _model.GeometryStore; }
+            get { return Model.GeometryStore; }
         }
 
         public IStepFileHeader Header
         {
-            get { return _model.Header; }
+            get { return Model.Header; }
         }
 
         public bool IsTransactional
         {
-            get { return _model.IsTransactional; }
+            get { return Model.IsTransactional; }
         }
 
         public string Location
         {
-            get => ModelProvider.GetLocation(_model);
+            get => ModelProvider.GetLocation(Model);
         }
 
         public IEntityCollection Instances
         {
-            get { return _model.Instances; }
+            get { return Model.Instances; }
+        }
+
+        /// <summary>
+        /// Returns a list of the handles to only the entities in this model
+        /// Note this do NOT include entities that are in any federated models
+        /// </summary>
+        public IList<XbimInstanceHandle> InstanceHandles
+        {
+            get { return Model.InstanceHandles.ToList(); }
         }
 
         bool IModel.Activate(IPersistEntity owningEntity)
         {
-            return _model.Activate(owningEntity);
+            return Model.Activate(owningEntity);
         }
 
         public void Delete(IPersistEntity entity)
         {
-            _model.Delete(entity);
+            Model.Delete(entity);
         }
 
         public ITransaction BeginTransaction(string name = null)
         {
 
-            if(_model.IsTransactional)
+            if(Model.IsTransactional)
             {
-                return _model.BeginTransaction(name);
+                return Model.BeginTransaction(name);
             }
             else
             {
                 throw new XbimException("Native store does not support transactions");
             }
-            //var esentModel = _model as EsentModel;
-            //if (esentModel != null) //we need to do transaction handling on Esent model, make sure we can write to it
-            //    return esentModel.BeginTransaction(name);
-
-            //var memoryModel = _model as MemoryModel;
-            //if (memoryModel == null)
-            //    throw new XbimException("Native store does not support transactions");
-            //return memoryModel.BeginTransaction(name);
         }
 
         public ITransaction CurrentTransaction
         {
-            get { return _model.CurrentTransaction; }
+            get { return Model.CurrentTransaction; }
         }
 
         public ExpressMetaData Metadata
         {
-            get { return _model.Metadata; }
+            get { return Model.Metadata; }
         }
 
         public IModelFactors ModelFactors
         {
-            get { return _model.ModelFactors; }
+            get { return Model.ModelFactors; }
         }
-
-        public string FileName { get; set; }
 
         public T InsertCopy<T>(T toCopy, XbimInstanceHandleMap mappings, PropertyTranformDelegate propTransform, bool includeInverses,
             bool keepLabels) where T : IPersistEntity
         {
-            return _model.InsertCopy(toCopy, mappings, propTransform, includeInverses, keepLabels);
+            return Model.InsertCopy(toCopy, mappings, propTransform, includeInverses, keepLabels);
         }
 
         public void ForEach<TSource>(IEnumerable<TSource> source, Action<TSource> body) where TSource : IPersistEntity
         {
-            _model.ForEach(source, body);
+            Model.ForEach(source, body);
         }
 
-        public void Dispose()
+        
+        public IEntityCache EntityCache => Model.EntityCache;
+
+        public IInverseCache BeginInverseCaching()
         {
-            Dispose(true);
-            // Take yourself off the Finalization queue 
-            // to prevent finalization code for this object
-            // from executing a second time.
-            GC.SuppressFinalize(this);
+            return Model.BeginInverseCaching();
         }
 
-        protected void Dispose(bool disposing)
+        public IEntityCache BeginEntityCaching()
         {
-            if (!_disposed)
-            {
-                try
-                {
-                    // If disposing equals true, dispose all managed 
-                    // and unmanaged resources.
-                    if (disposing)
-                    {
-                        Close();
-                        //release event handlers
-                        if (_model != null)
-                        {
-                            _model.EntityDeleted -= _model_EntityDeleted;
-                            _model.EntityNew -= _model_EntityNew;
-                            _model.EntityModified -= _model_EntityModified;
-                            //if(EditorDetails != null)
-                            //{
-                            //    _model.EntityNew -= IfcRootInit;
-                            //    _model.EntityModified -= IfcRootModified;
-                            //}
-                        }
-
-                        //managed resources
-                        if (_model is IDisposable disposeInterface)
-                            disposeInterface.Dispose();
-                    }
-                    //unmanaged, mostly Esent related                  
-                }
-                catch
-                {
-                    // ignored
-                }
-            }
-            _disposed = true;
+            return Model.BeginEntityCaching();
         }
+        public XbimSchemaVersion SchemaVersion => Model.SchemaVersion;
 
-
-
+        #endregion // IModel
         /// <summary>
         /// Closes the store and disposes of all resources. The store is invalid after this call
         /// </summary>
@@ -378,33 +351,27 @@ namespace Xbim.Ifc
             {
                 referencedModel.Close();
             }
-            ModelProvider.Close(_model);
+            ModelProvider.Close(Model);
 
-        }
-
-        /// <summary>
-        /// Creates a Database store at the specified location
-        /// </summary>
-        /// <param name="filePath"></param>
-        /// <param name="editorDetails"></param>
-        /// <param name="ifcVersion"></param>
-        /// <returns></returns>
-        public static IfcStore Create(string filePath, XbimEditorCredentials editorDetails, XbimSchemaVersion ifcVersion)
-        {
-            return new IfcStore(filePath, ifcVersion, editorDetails);
-        }
-
-        public static IfcStore Create(XbimEditorCredentials editorDetails, XbimSchemaVersion ifcVersion, XbimStoreType storageType)
-        {
-            return new IfcStore(storageType, ifcVersion, editorDetails);
-        }
-
-        public static IfcStore Create(XbimSchemaVersion ifcVersion, XbimStoreType storageType)
-        {
-            return new IfcStore(storageType, ifcVersion, null);
         }
 
         #region OwnerHistory Management
+
+
+        private void Model_EntityDeleted(IPersistEntity entity)
+        {
+            if (EntityDeleted != null) EntityDeleted.Invoke(entity);
+        }
+
+        private void Model_EntityNew(IPersistEntity entity)
+        {
+            if (EntityNew != null) EntityNew.Invoke(entity);
+        }
+
+        private void Model_EntityModified(IPersistEntity entity, int property)
+        {
+            if (EntityModified != null) EntityModified.Invoke(entity, property);
+        }
 
         private void IfcRootModified(IPersistEntity entity, int property)
         {
@@ -444,7 +411,7 @@ namespace Xbim.Ifc
                 if (EditorDetails == null)
                     return null;
 
-                if (_schema == XbimSchemaVersion.Ifc4 || _schema == XbimSchemaVersion.Ifc4x1)
+                if (SchemaVersion == XbimSchemaVersion.Ifc4 || SchemaVersion == XbimSchemaVersion.Ifc4x1)
                 {
                     var person = Instances.New<Ifc4.ActorResource.IfcPerson>(p =>
                     {
@@ -492,7 +459,7 @@ namespace Xbim.Ifc
                 if (EditorDetails == null)
                     return null;
 
-                if (_schema == XbimSchemaVersion.Ifc4 || _schema == XbimSchemaVersion.Ifc4x1)
+                if (SchemaVersion == XbimSchemaVersion.Ifc4 || SchemaVersion == XbimSchemaVersion.Ifc4x1)
                     return _defaultOwningApplication ??
                          (_defaultOwningApplication =
                              Instances.New<Ifc4.UtilityResource.IfcApplication>(a =>
@@ -524,7 +491,7 @@ namespace Xbim.Ifc
             {
                 if (_ownerHistoryAddObject != null)
                     return _ownerHistoryAddObject;
-                if (_schema == XbimSchemaVersion.Ifc4 || _schema == XbimSchemaVersion.Ifc4x1)
+                if (SchemaVersion == XbimSchemaVersion.Ifc4 || SchemaVersion == XbimSchemaVersion.Ifc4x1)
                 {
                     var histAdd = Instances.New<Ifc4.UtilityResource.IfcOwnerHistory>();
                     histAdd.OwningUser = (Ifc4.ActorResource.IfcPersonAndOrganization)DefaultOwningUser;
@@ -550,7 +517,7 @@ namespace Xbim.Ifc
             {
                 if (_ownerHistoryModifyObject != null)
                     return _ownerHistoryModifyObject;
-                if (_schema == XbimSchemaVersion.Ifc4 || _schema == XbimSchemaVersion.Ifc4x1)
+                if (SchemaVersion == XbimSchemaVersion.Ifc4 || SchemaVersion == XbimSchemaVersion.Ifc4x1)
                 {
                     var histmod = Instances.New<Ifc4.UtilityResource.IfcOwnerHistory>();
                     histmod.OwningUser = (Ifc4.ActorResource.IfcPersonAndOrganization)DefaultOwningUser;
@@ -648,7 +615,7 @@ namespace Xbim.Ifc
             FileName = actualFileName;
             if (actualFormat.HasFlag(StorageType.Xbim)) //special case for xbim
             {
-                ModelProvider.Persist(_model, actualFileName, progDelegate);
+                ModelProvider.Persist(Model, actualFileName, progDelegate);
             }
             else
             {
@@ -656,302 +623,15 @@ namespace Xbim.Ifc
                 {
                     if (actualFormat.HasFlag(StorageType.IfcZip))
                         //do zip first so that xml and ifc are not confused by the combination of flags
-                        SaveAsIfcZip(fileStream, Path.GetFileName(actualFileName), actualFormat, progDelegate);
+                        IfcStoreExportExtensions.SaveAsIfcZip(this, fileStream, Path.GetFileName(actualFileName), actualFormat, progDelegate);
                     else if (actualFormat.HasFlag(StorageType.Ifc))
-                        SaveAsIfc(fileStream, progDelegate);
+                        IfcStoreExportExtensions.SaveAsIfc(this, fileStream, progDelegate);
                     else if (actualFormat.HasFlag(StorageType.IfcXml))
-                        SaveAsIfcXml(fileStream, progDelegate);
+                        IfcStoreExportExtensions.SaveAsIfcXml(this, fileStream, progDelegate);
 
                 }
             }
         }
-
-        public void SaveAsIfcXml(Stream stream, ReportProgressDelegate progDelegate = null)
-        {
-            var settings = new XmlWriterSettings { Indent = true };
-            using (var xmlWriter = XmlWriter.Create(stream, settings))
-            {
-                if (_schema == XbimSchemaVersion.Ifc2X3)
-                {
-                    var writer = new IfcXmlWriter3();
-                    writer.Write(_model, xmlWriter, _model.Instances);
-
-                }
-                else if (_schema == XbimSchemaVersion.Ifc4 || _schema == XbimSchemaVersion.Ifc4x1)
-                {
-                    var writer = new XbimXmlWriter4(XbimXmlSettings.IFC4Add2);
-                    var project = _model.Instances.OfType<Ifc4.Kernel.IfcProject>();
-                    var products = _model.Instances.OfType<Ifc4.Kernel.IfcObject>();
-                    var relations = _model.Instances.OfType<Ifc4.Kernel.IfcRelationship>();
-
-                    var all =
-                        new IPersistEntity[] { }
-                        //start from root
-                            .Concat(project)
-                            //add all products not referenced in the project tree
-                            .Concat(products)
-                            //add all relations which are not inversed
-                            .Concat(relations)
-                            //make sure all other objects will get written
-                            .Concat(_model.Instances);
-
-                    writer.Write(_model, xmlWriter, all);
-                }
-                xmlWriter.Close();
-            }
-        }
-
-        public void SaveAsIfc(Stream stream, ReportProgressDelegate progDelegate = null)
-        {
-
-            using (TextWriter tw = new StreamWriter(stream))
-            {
-                Part21Writer.Write(_model, tw, _model.Metadata, null, progDelegate);
-                tw.Flush();
-            }
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="stream">The stream will be closed and flushed on return</param>
-        /// <param name="zipEntryName">The name of the file zipped inside the file</param>
-        /// <param name="storageType">Specify IfcZip and then either IfcXml or Ifc</param>
-        /// <param name="progDelegate"></param>
-        public void SaveAsIfcZip(Stream stream, string zipEntryName, StorageType storageType, ReportProgressDelegate progDelegate = null)
-        {
-            Debug.Assert(storageType.HasFlag(StorageType.IfcZip));
-            var fileBody = Path.ChangeExtension(zipEntryName,
-                storageType.HasFlag(StorageType.IfcXml) ? "ifcXml" : "ifc"
-                );
-
-            using (var zipStream = new ZipArchive(stream, ZipArchiveMode.Create))
-            {
-                var newEntry = zipStream.CreateEntry(fileBody);
-                using (var writer = newEntry.Open())
-                {
-
-                    if (storageType.HasFlag(StorageType.IfcXml))
-                        SaveAsIfcXml(writer, progDelegate);
-                    else //assume it is Ifc
-                        SaveAsIfc(writer, progDelegate);
-                }
-
-            }
-        }
-
-        /// <summary>
-        /// If translation is defined, returns matrix translated by the vector
-        /// </summary>
-        /// <param name="matrix">Input matrix</param>
-        /// <param name="translation">Translation</param>
-        /// <returns>Translated matrix</returns>
-        private XbimMatrix3D Translate(XbimMatrix3D matrix, IVector3D translation)
-        {
-            if (translation == null) return matrix;
-            var translationMatrix = XbimMatrix3D.CreateTranslation(translation.X, translation.Y, translation.Z);
-            return XbimMatrix3D.Multiply(matrix, translationMatrix);
-        }
-
-        /// <summary>
-        /// This function is used to generate the .wexbim model files.
-        /// </summary>
-        /// <param name="binaryStream">An open writable streamer.</param>
-        /// <param name="products">Optional products to be written to the wexBIM file. If null, all products from the model will be saved</param>
-        public void SaveAsWexBim(BinaryWriter binaryStream, IEnumerable<IIfcProduct> products = null, IVector3D translation = null)
-        {
-            products = products ?? Instances.OfType<IIfcProduct>();
-            // ReSharper disable RedundantCast
-            if (GeometryStore == null) throw new XbimException("Geometry store has not been initialised");
-            // ReSharper disable once CollectionNeverUpdated.Local
-            var colourMap = new XbimColourMap();
-            using (var geomRead = GeometryStore.BeginRead())
-            {
-
-                var lookup = geomRead.ShapeGeometries;
-                var styles = geomRead.StyleIds;
-                var regions = geomRead.ContextRegions.SelectMany(r => r).ToList();
-                //we need to get all the default styles for various products
-                var defaultStyles = geomRead.ShapeInstances.Select(i => -(int)i.IfcTypeId).Distinct();
-                var allStyles = defaultStyles.Concat(styles).ToList();
-                int numberOfGeometries = 0;
-                int numberOfVertices = 0;
-                int numberOfTriangles = 0;
-                int numberOfMatrices = 0;
-                int numberOfProducts = 0;
-                int numberOfStyles = allStyles.Count;
-                //start writing out
-
-                binaryStream.Write((Int32)WexBimId); //magic number
-
-                binaryStream.Write((byte)2); //version of stream, arrays now packed as doubles
-                var start = (int)binaryStream.Seek(0, SeekOrigin.Current);
-                binaryStream.Write((Int32)0); //number of shapes
-                binaryStream.Write((Int32)0); //number of vertices
-                binaryStream.Write((Int32)0); //number of triangles
-                binaryStream.Write((Int32)0); //number of matrices
-                binaryStream.Write((Int32)0); //number of products
-                binaryStream.Write((Int32)numberOfStyles); //number of styles
-                binaryStream.Write(Convert.ToSingle(_model.ModelFactors.OneMetre));
-                //write out conversion to meter factor
-
-                binaryStream.Write(Convert.ToInt16(regions.Count)); //write out the population data
-                var t = XbimMatrix3D.Identity;
-                t = Translate(t, translation);
-                foreach (var r in regions)
-                {
-                    binaryStream.Write((Int32)(r.Population));
-                    var bounds = r.ToXbimRect3D();
-                    var centre = t.Transform(r.Centre);
-                    //write out the centre of the region
-                    binaryStream.Write((Single)centre.X);
-                    binaryStream.Write((Single)centre.Y);
-                    binaryStream.Write((Single)centre.Z);
-                    //bounding box of largest region
-                    binaryStream.Write(bounds.ToFloatArray());
-                }
-                //textures
-
-                foreach (var styleId in allStyles)
-                {
-                    XbimColour colour;
-                    if (styleId > 0)
-                    {
-                        var ss = (IIfcSurfaceStyle)Instances[styleId];
-                        var texture = XbimTexture.Create(ss);
-                        colour = texture.ColourMap.FirstOrDefault();
-                    }
-                    else //use the default in the colour map for the enetity type
-                    {
-                        var theType = _model.Metadata.GetType((short)Math.Abs(styleId));
-                        colour = colourMap[theType.Name];
-                    }
-                    if (colour == null) colour = XbimColour.DefaultColour;
-                    binaryStream.Write((Int32)styleId); //style ID                       
-                    binaryStream.Write((Single)colour.Red);
-                    binaryStream.Write((Single)colour.Green);
-                    binaryStream.Write((Single)colour.Blue);
-                    binaryStream.Write((Single)colour.Alpha);
-
-                }
-
-                //write out all the product bounding boxes
-                var prodIds = new HashSet<int>();
-                foreach (var product in products)
-                {
-                    if (product is IIfcFeatureElement) continue;
-                    prodIds.Add(product.EntityLabel);
-
-                    var bb = XbimRect3D.Empty;
-                    foreach (var si in geomRead.ShapeInstancesOfEntity(product))
-                    {
-                        var transformation = Translate(si.Transformation, translation);
-                        var bbPart = XbimRect3D.TransformBy(si.BoundingBox, transformation);
-                        //make sure we put the box in the right place and then convert to axis aligned
-                        if (bb.IsEmpty) bb = bbPart;
-                        else
-                            bb.Union(bbPart);
-                    }
-                    //do not write out anything with no geometry
-                    if (bb.IsEmpty) continue;
-
-                    binaryStream.Write((Int32)product.EntityLabel);
-                    binaryStream.Write((UInt16)_model.Metadata.ExpressTypeId(product));
-                    binaryStream.Write(bb.ToFloatArray());
-                    numberOfProducts++;
-                }
-
-                //projections and openings have already been applied, 
-
-                var toIgnore = new short[4];
-                toIgnore[0] = _model.Metadata.ExpressTypeId("IFCOPENINGELEMENT");
-                toIgnore[1] = _model.Metadata.ExpressTypeId("IFCPROJECTIONELEMENT");
-                if (SchemaVersion == XbimSchemaVersion.Ifc4 || _schema == XbimSchemaVersion.Ifc4x1)
-                {
-                    toIgnore[2] = _model.Metadata.ExpressTypeId("IFCVOIDINGFEATURE");
-                    toIgnore[3] = _model.Metadata.ExpressTypeId("IFCSURFACEFEATURE");
-                }
-
-                foreach (var geometry in lookup)
-                {
-                    if (geometry.ShapeData.Length <= 0) //no geometry to display so don't write out any products for it
-                        continue;
-                    var instances = geomRead.ShapeInstancesOfGeometry(geometry.ShapeLabel);
-
-
-
-                    var xbimShapeInstances = instances.Where(si => !toIgnore.Contains(si.IfcTypeId) &&
-                                                                 si.RepresentationType ==
-                                                                 XbimGeometryRepresentationType
-                                                                     .OpeningsAndAdditionsIncluded && prodIds.Contains(si.IfcProductLabel)).ToList();
-                    if (!xbimShapeInstances.Any()) continue;
-                    numberOfGeometries++;
-                    binaryStream.Write(xbimShapeInstances.Count); //the number of repetitions of the geometry
-                    if (xbimShapeInstances.Count > 1)
-                    {
-                        foreach (IXbimShapeInstanceData xbimShapeInstance in xbimShapeInstances)
-                        //write out each of the ids style and transforms
-                        {
-                            binaryStream.Write(xbimShapeInstance.IfcProductLabel);
-                            binaryStream.Write((UInt16)xbimShapeInstance.IfcTypeId);
-                            binaryStream.Write((UInt32)xbimShapeInstance.InstanceLabel);
-                            binaryStream.Write((Int32)xbimShapeInstance.StyleLabel > 0
-                                ? xbimShapeInstance.StyleLabel
-                                : xbimShapeInstance.IfcTypeId * -1);
-
-                            var transformation = Translate(XbimMatrix3D.FromArray(xbimShapeInstance.Transformation), translation);
-                            binaryStream.Write(transformation.ToArray());
-                            numberOfTriangles +=
-                                XbimShapeTriangulation.TriangleCount(((IXbimShapeGeometryData)geometry).ShapeData);
-                            numberOfMatrices++;
-                        }
-                        numberOfVertices +=
-                            XbimShapeTriangulation.VerticesCount(((IXbimShapeGeometryData)geometry).ShapeData);
-                        // binaryStream.Write(geometry.ShapeData);
-                        var ms = new MemoryStream(((IXbimShapeGeometryData)geometry).ShapeData);
-                        var br = new BinaryReader(ms);
-                        var tr = br.ReadShapeTriangulation();
-
-                        tr.Write(binaryStream);
-                    }
-                    else //now do the single instances
-                    {
-                        var xbimShapeInstance = xbimShapeInstances[0];
-
-                        // IXbimShapeGeometryData geometry = ShapeGeometry(kv.Key);
-                        binaryStream.Write((Int32)xbimShapeInstance.IfcProductLabel);
-                        binaryStream.Write((UInt16)xbimShapeInstance.IfcTypeId);
-                        binaryStream.Write((Int32)xbimShapeInstance.InstanceLabel);
-                        binaryStream.Write((Int32)xbimShapeInstance.StyleLabel > 0
-                            ? xbimShapeInstance.StyleLabel
-                            : xbimShapeInstance.IfcTypeId * -1);
-
-                        //Read all vertices and normals in the geometry stream and transform
-
-                        var ms = new MemoryStream(((IXbimShapeGeometryData)geometry).ShapeData);
-                        var br = new BinaryReader(ms);
-                        var tr = br.ReadShapeTriangulation();
-                        var transformation = Translate(xbimShapeInstance.Transformation, translation);
-                        var trTransformed = tr.Transform(transformation);
-                        trTransformed.Write(binaryStream);
-                        numberOfTriangles += XbimShapeTriangulation.TriangleCount(((IXbimShapeGeometryData)geometry).ShapeData);
-                        numberOfVertices += XbimShapeTriangulation.VerticesCount(((IXbimShapeGeometryData)geometry).ShapeData);
-                    }
-                }
-
-
-                binaryStream.Seek(start, SeekOrigin.Begin);
-                binaryStream.Write((Int32)numberOfGeometries);
-                binaryStream.Write((Int32)numberOfVertices);
-                binaryStream.Write((Int32)numberOfTriangles);
-                binaryStream.Write((Int32)numberOfMatrices);
-                binaryStream.Write((Int32)numberOfProducts);
-                binaryStream.Seek(0, SeekOrigin.End); //go back to end
-                // ReSharper restore RedundantCast
-            }
-        }
-
-        public const int WexBimId = 94132117;
 
         /// <summary>
         /// Calculates and sets the model factors, call every time a unit of measurement is changed
@@ -1066,8 +746,9 @@ namespace Xbim.Ifc
             }
         }
 
-        #region Referenced Models functions
+        #region Referenced Models functions / Federation
 
+     
         /// <summary>
         /// Adds a model as a reference or federated model, do not call inside a transaction
         /// </summary>
@@ -1080,7 +761,7 @@ namespace Xbim.Ifc
             XbimReferencedModel retVal;
             using (var txn = BeginTransaction())
             {
-                if (_schema == XbimSchemaVersion.Ifc4 || _schema == XbimSchemaVersion.Ifc4x1)
+                if (SchemaVersion == XbimSchemaVersion.Ifc4 || SchemaVersion == XbimSchemaVersion.Ifc4x1)
                 {
                     var role = Instances.New<Ifc4.ActorResource.IfcActorRole>();
                     role.RoleString = organisationRole;
@@ -1211,9 +892,7 @@ namespace Xbim.Ifc
             }
         }
 
-        #endregion
-        #region Federation
-
+        public string FileName { get; set; }
         public IEnumerable<IReferencedModel> ReferencedModels
         {
             get { return _referencedModels.AsEnumerable(); }
@@ -1232,21 +911,6 @@ namespace Xbim.Ifc
             get { return _referencedModels.Any(); }
         }
 
-        ///// <summary>
-        ///// Returns an enumerable of the handles to all entities in the model
-        ///// Note this includes entities that are in any federated models
-        ///// </summary>
-        //public IEnumerable<XbimInstanceHandle> AllInstancesHandles
-        //{
-        //    get
-        //    {
-        //        foreach (var h in InstanceHandles)
-        //            yield return h;
-        //        foreach (var refModel in ReferencedModels.Where(r => r.Model is EsentModel).Select(r => r.Model as EsentModel))
-        //            foreach (var h in refModel.AllInstancesHandles)
-        //                yield return h;
-        //    }
-        //}
 
         public void EnsureUniqueUserDefinedId()
         {
@@ -1258,11 +922,11 @@ namespace Xbim.Ifc
                 model.UserDefinedId = iId++;
             }
         }
-        #endregion
+
 
         public IModel ReferencingModel
         {
-            get { return _model; }
+            get { return Model; }
         }
 
         public IReadOnlyEntityCollection FederatedInstances
@@ -1278,245 +942,6 @@ namespace Xbim.Ifc
                 return allModels.SelectMany(m => m.InstanceHandles).ToList();
             }
         }
-
-        /// <summary>
-        /// Returns a list of the handles to only the entities in this model
-        /// Note this do NOT include entities that are in any federated models
-        /// </summary>
-
-        public IList<XbimInstanceHandle> InstanceHandles
-        {
-            get { return _model.InstanceHandles.ToList(); }
-        }
-
-        public ILogger Logger { get => _model.Logger; set => _model.Logger = value; }
-
-        public IEntityCache EntityCache => _model.EntityCache;
-
-        XbimSchemaVersion SchemaVersion => _model.SchemaVersion;
-
-        XbimSchemaVersion IModel.SchemaVersion => throw new NotImplementedException();
-
-        public XbimEditorCredentials EditorDetails { get ; private set; }
-
-        #region Insert products with context
-
-        private List<IIfcProduct> _primaryElements = new List<IIfcProduct>();
-        private readonly List<IIfcProduct> _decomposition = new List<IIfcProduct>();
-        private bool _includeGeometry;
-
-        /// <summary>
-        /// This is a higher level function which uses InsertCopy function alongside with the knowledge of IFC schema to copy over
-        /// products with their types and other related information (classification, aggregation, documents, properties) and optionally
-        /// geometry. It will also bring in spatial hierarchy relevant to selected products. However, resulting model is not guaranteed 
-        /// to be compliant with any Model View Definition unless you explicitly check the compliance. Context of a single product tend to 
-        /// consist from hundreds of objects which need to be identified and copied over so this operation might be potentially expensive.
-        /// You should never call this function more than once between two models. It not only selects objects to be copied over but also
-        /// excludes other objects from being copied over so that it doesn't bring the entire model in a chain dependencies. This means
-        /// that some objects are modified (like spatial relations) and won't get updated which would lead to an inconsistent copy.
-        /// </summary>
-        /// <param name="products">Products from other model to be inserted into this model</param>
-        /// <param name="includeGeometry">If TRUE, geometry of the products will be copied over.</param>
-        /// <param name="keepLabels">If TRUE, entity labels from original model will be used. Always set this to FALSE
-        /// if you are going to insert products from multiple source models or if you are going to insert products to a non-empty model</param>
-        /// <param name="mappings">Mappings to avoid multiple insertion of objects. Keep a single instance for insertion between two models.
-        /// If you also use InsertCopy() function for some other insertions, use the same instance of mappings.</param>
-        public void InsertCopy(IEnumerable<IIfcProduct> products, bool includeGeometry, bool keepLabels, XbimInstanceHandleMap mappings)
-        {
-            _primaryElements.Clear();
-            _decomposition.Clear();
-            _includeGeometry = includeGeometry;
-
-            var roots = products.Cast<IPersistEntity>().ToList();
-            //return if there is nothing to insert
-            if (!roots.Any())
-                return;
-
-            var source = roots.First().Model;
-            if (source == this)
-                //don't do anything if the source and target are the same
-                return;
-
-            var toInsert = GetEntitiesToInsert(source, roots);
-            //create new cache is none is defined
-            var cache = mappings ?? new XbimInstanceHandleMap(source, this);
-
-            foreach (var entity in toInsert)
-                InsertCopy(entity, cache, Filter, true, keepLabels);
-        }
-
-        private IEnumerable<IPersistEntity> GetEntitiesToInsert(IModel model, List<IPersistEntity> roots)
-        {
-            _primaryElements = roots.OfType<IIfcProduct>().ToList();
-
-            //add any aggregated elements. For example IfcRoof is typically aggregation of one or more slabs so we need to bring
-            //them along to have all the information both for geometry and for properties and materials.
-            //This has to happen before we add spatial hierarchy or it would bring in full hierarchy which is not an intention
-            var decompositionRels = GetAggregations(_primaryElements.ToList(), model).ToList();
-            _primaryElements.AddRange(_decomposition);
-            roots.AddRange(decompositionRels);
-
-            //we should add spatial hierarchy right here so it brings its attributes as well
-            var spatialRels = model.Instances.Where<IIfcRelContainedInSpatialStructure>(
-                r => _primaryElements.Any(e => r.RelatedElements.Contains(e))).ToList();
-            var spatialRefs =
-                model.Instances.Where<IIfcRelReferencedInSpatialStructure>(
-                    r => _primaryElements.Any(e => r.RelatedElements.Contains(e))).ToList();
-            var bottomSpatialHierarchy =
-                spatialRels.Select(r => r.RelatingStructure).Union(spatialRefs.Select(r => r.RelatingStructure)).ToList();
-            var spatialAggregations = GetUpstreamHierarchy(bottomSpatialHierarchy, model).ToList();
-
-            //add all spatial elements from bottom and from upstream hierarchy
-            _primaryElements.AddRange(bottomSpatialHierarchy);
-            _primaryElements.AddRange(spatialAggregations.Select(r => r.RelatingObject).OfType<IIfcProduct>());
-            roots.AddRange(spatialAggregations);
-            roots.AddRange(spatialRels);
-            roots.AddRange(spatialRefs);
-
-            //we should add any feature elements used to subtract mass from a product
-            var featureRels = GetFeatureRelations(_primaryElements).ToList();
-            var openings = featureRels.Select(r => r.RelatedOpeningElement);
-            _primaryElements.AddRange(openings);
-            roots.AddRange(featureRels);
-
-            //object types and properties for all primary products (elements and spatial elements)
-            roots.AddRange(_primaryElements.SelectMany(p => p.IsDefinedBy));
-            roots.AddRange(_primaryElements.SelectMany(p => p.IsTypedBy));
-
-
-
-            //assignmnet to groups will bring in all system aggregarions if defined in the file
-            roots.AddRange(_primaryElements.SelectMany(p => p.HasAssignments));
-
-            //associations with classification, material and documents
-            roots.AddRange(_primaryElements.SelectMany(p => p.HasAssociations));
-
-            return roots;
-        }
-
-        private object Filter(ExpressMetaProperty property, object parentObject)
-        {
-            //ignore inverses except for style
-            if (property.IsInverse)
-                return property.Name == "StyledByItem" ? property.PropertyInfo.GetValue(parentObject, null) : null;
-
-            if (_primaryElements != null && _primaryElements.Any())
-            {
-                if (typeof(IIfcProduct).IsAssignableFrom(property.PropertyInfo.PropertyType))
-                {
-                    var element = property.PropertyInfo.GetValue(parentObject, null) as IIfcProduct;
-                    if (element != null && _primaryElements.Contains(element))
-                        return element;
-                    return null;
-                }
-                if (property.EnumerableType != null && !property.EnumerableType.IsValueType && property.EnumerableType != typeof(string))
-                {
-                    //this can either be a list of IPersistEntity or select type. The very base type is IPersist
-                    var entities = property.PropertyInfo.GetValue(parentObject, null) as IEnumerable<IPersist>;
-                    if (entities != null)
-                    {
-                        var persistEntities = entities as IList<IPersist> ?? entities.ToList();
-                        var elementsToRemove = persistEntities.OfType<IIfcProduct>().Where(e => !_primaryElements.Contains(e)).ToList();
-                        //if there are no IfcElements return what is in there with no care
-                        if (elementsToRemove.Any())
-                            //return original values excluding elements not included in the primary set
-                            return persistEntities.Except(elementsToRemove).ToList();
-                    }
-                }
-            }
-
-            //if geometry is to be included don't filter it out
-            if (_includeGeometry)
-                return property.PropertyInfo.GetValue(parentObject, null);
-
-            //leave out geometry and placement of products
-            if (parentObject is IIfcProduct &&
-                (property.PropertyInfo.Name == "Representation" || property.PropertyInfo.Name == "ObjectPlacement")
-                )
-                return null;
-
-            //leave out representation maps
-            if (parentObject is IIfcTypeProduct && property.PropertyInfo.Name == "RepresentationMaps")
-                return null;
-
-            //leave out eventual connection geometry
-            if (parentObject is IIfcRelSpaceBoundary && property.PropertyInfo.Name == "ConnectionGeometry")
-                return null;
-
-            //return the value for anything else
-            return property.PropertyInfo.GetValue(parentObject, null);
-        }
-
-        private static IEnumerable<IIfcRelVoidsElement> GetFeatureRelations(IEnumerable<IIfcProduct> products)
-        {
-            var elements = products.OfType<IIfcElement>().ToList();
-            if (!elements.Any()) yield break;
-            var model = elements.First().Model;
-            var rels = model.Instances.Where<IIfcRelVoidsElement>(r => elements.Any(e => Equals(e, r.RelatingBuildingElement)));
-            foreach (var rel in rels)
-                yield return rel;
-        }
-
-        private IEnumerable<IIfcRelDecomposes> GetAggregations(List<IIfcProduct> products, IModel model)
-        {
-            _decomposition.Clear();
-            while (true)
-            {
-                if (!products.Any())
-                    yield break;
-
-                var products1 = products;
-                var rels = model.Instances.Where<IIfcRelDecomposes>(r =>
-                {
-                    if (r is IIfcRelAggregates aggr)
-                        return products1.Any(p => Equals(aggr.RelatingObject, p));
-                    if (r is IIfcRelNests nest)
-                        return products1.Any(p => Equals(nest.RelatingObject, p));
-                    if (r is IIfcRelProjectsElement prj)
-                        return products1.Any(p => Equals(prj.RelatingElement, p));
-                    if (r is IIfcRelVoidsElement voids)
-                        return products1.Any(p => Equals(voids.RelatingBuildingElement, p));
-                    return false;
-
-                }).ToList();
-                var relatedProducts = rels.SelectMany(r =>
-                {
-                    if (r is IIfcRelAggregates aggr)
-                        return aggr.RelatedObjects.OfType<IIfcProduct>();
-                    if (r is IIfcRelNests nest)
-                        return nest.RelatedObjects.OfType<IIfcProduct>();
-                    if (r is IIfcRelProjectsElement prj)
-                        return new IIfcProduct[] { prj.RelatedFeatureElement };
-                    if (r is IIfcRelVoidsElement voids)
-                        return new IIfcProduct[] { voids.RelatedOpeningElement };
-                    return null;
-                }).Where(p => p != null).ToList();
-
-                foreach (var rel in rels)
-                    yield return rel;
-
-                products = relatedProducts;
-                _decomposition.AddRange(products);
-            }
-        }
-
-        private static IEnumerable<IIfcRelAggregates> GetUpstreamHierarchy(IEnumerable<IIfcSpatialElement> spatialStructureElements, IModel model)
-        {
-            while (true)
-            {
-                var elements = spatialStructureElements.ToList();
-                if (!elements.Any())
-                    yield break;
-
-                var rels = model.Instances.Where<IIfcRelAggregates>(r => elements.Any(s => r.RelatedObjects.Contains(s))).ToList();
-                var decomposing = rels.Select(r => r.RelatingObject).OfType<IIfcSpatialStructureElement>();
-
-                foreach (var rel in rels)
-                    yield return rel;
-
-                spatialStructureElements = decomposing;
-            }
-        }
         #endregion
 
         #region Equality
@@ -1528,7 +953,7 @@ namespace Xbim.Ifc
         /// <returns></returns>
         public bool Equals(IModel other)
         {
-            return ReferenceEquals(this, other) || ReferenceEquals(other, _model);
+            return ReferenceEquals(this, other) || ReferenceEquals(other, Model);
         }
 
         /// <summary>
@@ -1538,22 +963,12 @@ namespace Xbim.Ifc
         /// <returns></returns>
         public override bool Equals(object obj)
         {
-            return _model.Equals(obj) || ReferenceEquals(this, obj);
+            return Model.Equals(obj) || ReferenceEquals(this, obj);
         }
 
         public override int GetHashCode()
         {
-            return _model.GetHashCode();
-        }
-
-        public IInverseCache BeginInverseCaching()
-        {
-            return _model.BeginInverseCaching();
-        }
-
-        public IEntityCache BeginEntityCaching()
-        {
-            return _model.BeginEntityCaching();
+            return Model.GetHashCode();
         }
 
         public static bool operator == (IfcStore store, IModel model)
@@ -1565,7 +980,7 @@ namespace Xbim.Ifc
             if (ReferenceEquals(model, null))
                 return false;
 
-            return store._model.Equals(model);
+            return store.Model.Equals(model);
         }
 
         public static bool operator !=(IfcStore store, IModel model)
@@ -1573,6 +988,55 @@ namespace Xbim.Ifc
             return !(store == model);
         }
 
+        #endregion
+
+        #region Dispose
+        public void Dispose()
+        {
+            Dispose(true);
+            // Take yourself off the Finalization queue 
+            // to prevent finalization code for this object
+            // from executing a second time.
+            GC.SuppressFinalize(this);
+        }
+
+        protected void Dispose(bool disposing)
+        {
+            if (!_disposed)
+            {
+                try
+                {
+                    // If disposing equals true, dispose all managed 
+                    // and unmanaged resources.
+                    if (disposing)
+                    {
+                        Close();
+                        //release event handlers
+                        if (Model != null)
+                        {
+                            Model.EntityDeleted -= Model_EntityDeleted;
+                            Model.EntityNew -= Model_EntityNew;
+                            Model.EntityModified -= Model_EntityModified;
+                            if (EditorDetails != null)
+                            {
+                                Model.EntityNew -= IfcRootInit;
+                                Model.EntityModified -= IfcRootModified;
+                            }
+                        }
+
+                        //managed resources
+                        if (Model is IDisposable disposeInterface)
+                            disposeInterface.Dispose();
+                    }
+                    //unmanaged, mostly Esent related                  
+                }
+                catch
+                {
+                    // ignored
+                }
+            }
+            _disposed = true;
+        }
         #endregion
     }
 }
